+++ x/evm/keeper/params_test.go
@@ -33,7 +33,8 @@ func (suite *KeeperTestSuite) TestParams() {
 				return params.EvmDenom
 			},
 			func() interface{} {
-				return suite.app.EvmKeeper.GetEVMDenom(suite.ctx)
+				evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
+				return evmParams.GetEvmDenom()
 			},
 			true,
 		},
@@ -45,7 +46,8 @@ func (suite *KeeperTestSuite) TestParams() {
 				return params.EnableCreate
 			},
 			func() interface{} {
-				return suite.app.EvmKeeper.GetEnableCreate(suite.ctx)
+				evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
+				return evmParams.GetEnableCreate()
 			},
 			true,
 		},
@@ -57,7 +59,8 @@ func (suite *KeeperTestSuite) TestParams() {
 				return params.EnableCall
 			},
 			func() interface{} {
-				return suite.app.EvmKeeper.GetEnableCall(suite.ctx)
+				evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
+				return evmParams.GetEnableCall()
 			},
 			true,
 		},
@@ -69,7 +72,8 @@ func (suite *KeeperTestSuite) TestParams() {
 				return params.AllowUnprotectedTxs
 			},
 			func() interface{} {
-				return suite.app.EvmKeeper.GetAllowUnprotectedTxs(suite.ctx)
+				evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
+				return evmParams.GetAllowUnprotectedTxs()
 			},
 			true,
 		},
@@ -81,7 +85,8 @@ func (suite *KeeperTestSuite) TestParams() {
 				return params.ChainConfig
 			},
 			func() interface{} {
-				return suite.app.EvmKeeper.GetChainConfig(suite.ctx)
+				evmParams := suite.app.EvmKeeper.GetParams(suite.ctx)
+				return evmParams.GetChainConfig()
 			},
 			true,
 		},