+++ x/evm/migrations/v4/migrate_test.go
@@ -3,15 +3,15 @@ package v4_test
 import (
 	"testing"
 
-	"github.com/evmos/ethermint/x/evm/types"
-	gogotypes "github.com/gogo/protobuf/types"
+	"github.com/stretchr/testify/require"
 
 	"github.com/cosmos/cosmos-sdk/testutil"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/evmos/ethermint/app"
 	"github.com/evmos/ethermint/encoding"
 	v4 "github.com/evmos/ethermint/x/evm/migrations/v4"
-	"github.com/stretchr/testify/require"
+	"github.com/evmos/ethermint/x/evm/types"
 )
 
 type mockSubspace struct {
@@ -31,34 +31,16 @@ func TestMigrate(t *testing.T) {
 	cdc := encCfg.Codec
 
 	storeKey := sdk.NewKVStoreKey(types.ModuleName)
-	tKey := sdk.NewTransientStoreKey("transient_test")
+	tKey := sdk.NewTransientStoreKey(types.TransientKey)
 	ctx := testutil.DefaultContext(storeKey, tKey)
 	kvStore := ctx.KVStore(storeKey)
 
 	legacySubspace := newMockSubspace(types.DefaultParams())
 	require.NoError(t, v4.MigrateStore(ctx, storeKey, legacySubspace, cdc))
 
-	// Get all the new parameters from the kvStore
-	var evmDenom string
-	bz := kvStore.Get(types.ParamStoreKeyEVMDenom)
-	evmDenom = string(bz)
+	paramsBz := kvStore.Get(types.KeyPrefixParams)
+	var params types.Params
+	cdc.MustUnmarshal(paramsBz, &params)
 
-	var allowUnprotectedTx gogotypes.BoolValue
-	bz = kvStore.Get(types.ParamStoreKeyAllowUnprotectedTxs)
-	cdc.MustUnmarshal(bz, &allowUnprotectedTx)
-
-	enableCreate := kvStore.Has(types.ParamStoreKeyEnableCreate)
-	enableCall := kvStore.Has(types.ParamStoreKeyEnableCall)
-
-	var chainCfg types.ChainConfig
-	bz = kvStore.Get(types.ParamStoreKeyChainConfig)
-	cdc.MustUnmarshal(bz, &chainCfg)
-
-	var extraEIPs types.ExtraEIPs
-	bz = kvStore.Get(types.ParamStoreKeyExtraEIPs)
-	cdc.MustUnmarshal(bz, &extraEIPs)
-	require.Equal(t, types.AvailableExtraEIPs, extraEIPs.EIPs)
-
-	params := types.NewParams(evmDenom, allowUnprotectedTx.Value, enableCreate, enableCall, chainCfg, extraEIPs)
-	require.Equal(t, legacySubspace.ps, params)
+	require.Equal(t, params, legacySubspace.ps)
 }